requirements.txt
strands-agents
strands-agents-tools
requests
pytest
